requirements.txt
Flask==3.1.1
flask_sqlalchemy==3.1.1
plotly==6.2.0
python-dotenv==1.1.1
SQLAlchemy==2.0.41

# pipreqs . --encoding utf-8 --ignore .venv,bin,lib,include
# pipreqs . --force --encoding utf-8 --ignore .venv,bin,lib,include
